Split form into separate components
